🏗️ Move and rename `feed` slice to `post` feature folder
In the new feature-based architecture, each feature has its own dedicated `store` subfolder. Slices are the final part migrated into it. To preserve Git history, moving and renaming this file has been done before applying deeper modifications. In fact, `postSlice` will be split for each `feed` sub-feature in next commits.
