+++ src/features/activity-feed/post/components/ActivityFeed/index.jsx
@@ -4,7 +4,7 @@ import { Box, Grid, CircularProgress } from '@mui/material'
 import { PostIdProvider } from '../../contexts/PostIdProvider'
 import { selectPosts, selectActivityFeedHasMorePosts, selectPostLoading } from '../../store/postsSelectors'
 import { fetchPosts } from '../../store/postsThunks'
-import { cleanFeedState } from '../../../../../redux/reducers/feed'
+import { cleanFeedState } from '../../store/postsSlice'
 import { ProfileUserIdContext } from '../../contexts/ProfileUserIdProvider'
 import { ActivityFeedHeader } from '../ActivityFeedHeader'
 import { PostForm } from '../PostForm'
